+++ src/felupe/mechanics/__init__.py
@@ -1,3 +1,4 @@
+from ._contact import ContactRigidPlane
 from ._curve import CharacteristicCurve
 from ._free_vibration import FreeVibration
 from ._helpers import Assemble, Evaluate, Results, StateNearlyIncompressible
@@ -24,6 +25,7 @@ __all__ = [
     "Job",
     "PointLoad",
     "Results",
+    "ContactRigidPlane",
     "SolidBody",
     "SolidBodyCauchyStress",
     "SolidBodyForce",